third_party/WebKit/LayoutTests/fast/events/ime-composition-events-segments-expected.txt

This tests that DOM L3 CompositionEvent segment related properties.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Case 1: Test from empty input field.
PASS event.type is "compositionstart"
PASS event.data is ""
PASS event.activeSegmentStart is 0
PASS event.activeSegmentEnd is 3
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 3
PASS event.type is "compositionupdate"
PASS event.data is "abcdefghi"
PASS event.activeSegmentStart is 0
PASS event.activeSegmentEnd is 3
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 3
PASS event.type is "compositionupdate"
PASS event.data is "abcdefghi"
PASS event.activeSegmentStart is 3
PASS event.activeSegmentEnd is 6
PASS segments.length is 3
PASS segments[i] is 0
PASS segments[i] is 3
PASS segments[i] is 6
PASS event.type is "compositionupdate"
PASS event.data is "abcdefghi"
PASS event.activeSegmentStart is 6
PASS event.activeSegmentEnd is 9
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 6
PASS event.type is "compositionend"
PASS event.data is "abcdefghi"
PASS event.activeSegmentStart is 6
PASS event.activeSegmentEnd is 9
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 6
PASS test.value is "abcdefghi"
Case 2: Test from non-empty input field.
PASS event.type is "compositionstart"
PASS event.data is ""
PASS event.activeSegmentStart is 0
PASS event.activeSegmentEnd is 3
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 3
PASS event.type is "compositionupdate"
PASS event.data is "123456789"
PASS event.activeSegmentStart is 0
PASS event.activeSegmentEnd is 3
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 3
PASS event.type is "compositionupdate"
PASS event.data is "123456789"
PASS event.activeSegmentStart is 3
PASS event.activeSegmentEnd is 6
PASS segments.length is 3
PASS segments[i] is 0
PASS segments[i] is 3
PASS segments[i] is 6
PASS event.type is "compositionupdate"
PASS event.data is "123456789"
PASS event.activeSegmentStart is 6
PASS event.activeSegmentEnd is 9
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 6
PASS event.type is "compositionend"
PASS event.data is "123456789"
PASS event.activeSegmentStart is 6
PASS event.activeSegmentEnd is 9
PASS segments.length is 2
PASS segments[i] is 0
PASS segments[i] is 6
PASS test.value is "abcdefghi123456789"
PASS successfullyParsed is true

TEST COMPLETE
